Adding OJ: A JavaScript library to build websites with objects
Docs: http://ojjs.org
Repo: http://github.com/ojjs/oj
Download: http://ojjs.org/download.html

In addition this includes the OJ plugins which extend the behavior
of OJ. I know it was asked to keep things lowercase, but
unfortunately these plugin names are case sensitive to match the
JS convention that Objects are UpperCase, and functions are lowerCase.
OJ can be extended with either so the capitalization helps you know
how to use it.

Let me know if you have questions, or if you would prefer the
oj plugins to be in a sub directory (they are an essential part
of OJ, but it is harder to measure their popularity as people
only star the main one).

Here is an example of a plugin in use:
http://jsfiddle.net/evanmoran/yPXuL/