Added Fluidbox
Fluidbox is a jQuery plugin that mimics the fluid-like lightbox module
offered on Medium. It started as a personal project but soon expanded
into a full-fledge jQuery plugin.

Fluidbox has, at the time of which this pull request is made, 37
watchers and 667 stars. It is available at:
https://github.com/terrymun/Fluidbox

Fluidbox depends on jQuery and Desandro’s imagesLoaded plugin
(https://github.com/desandro/imagesloaded).