Add jquery-ui-bootstrap
Source:
https://github.com/jquery-ui-bootstrap/jquery-ui-bootstrap/archive/v0.5p
re.zip

This is a somewhat scrappy-looking library with various version numbers
in filenames. There’s no simple way to pare them back, so I’m leaving
them as-is. Also the version number triggers an npm failure on my
computer, so this might have to be pulled if it breaks Travis.

Cross your fingers, everyone.