Updating Library AmMaps to 3.13.0
### AmMaps

URL: http://www.amcharts.com/javascript-maps/

The above library is free to use. Please check the section "Free even
for commercial use" on the above Library's URL

Since this library is not hosted on github, auto update has not been
configured.

Removed examples, .md, .txt files from the original library as this is
not required to be hosted on cdn.

Hopefully all is OK with this PR.

@PeterDaveHello. Sending PR again as requested. Could you please verify
the same.